# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem pencatatan absensi sholat siswa",
        "title": "API Sistem Absensi Sholat",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Login untuk sistem absensi menggunakan identifier (NIS untuk siswa, username untuk admin/guru) dan password. Mengembalikan data profil pengguna jika login berhasil",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login Pengguna",
                "parameters": [
                    {
                        "description": "Kredensial login (identifier dan password)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login berhasil - Mengembalikan profil lengkap pengguna beserta token",
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginResponseData"
                        }
                    },
                    "400": {
                        "description": "Permintaan tidak valid - Identifier atau password kosong, atau format JSON salah",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Tidak terotentikasi - Identifier tidak ditemukan atau password salah",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal - Database error atau sistem error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan data profil user yang sedang login (siswa atau staff)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get Current User",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data profil",
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Tidak terotentikasi - Token tidak valid atau tidak ada",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal - Database error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Mendaftarkan akun siswa baru untuk sistem absensi. Siswa harus memiliki NIS yang terdaftar di sistem dan menggunakan email Google",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Pendaftaran Siswa",
                "parameters": [
                    {
                        "description": "Data pendaftaran siswa",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pendaftaran berhasil - Akun siswa berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Permintaan tidak valid - NIS/Password/Email kosong, format email salah, atau bukan email Google",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "NIS tidak terdaftar - NIS tidak ditemukan di database siswa",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Akun sudah terdaftar - NIS sudah memiliki akun login",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal - Database error atau sistem error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.LoginRequest": {
            "type": "object",
            "required": [
                "identifier",
                "password"
            ],
            "properties": {
                "identifier": {
                    "description": "NIS for siswa, username for admin/guru",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handlers.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id_staff": {
                    "type": "integer"
                },
                "is_google_acct": {
                    "type": "boolean"
                },
                "jk": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "kelas": {
                    "type": "string"
                },
                "nama_siswa": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nip": {
                    "type": "string"
                },
                "nis": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "description": "Staff fields",
                    "type": "string"
                }
            }
        },
        "handlers.LoginResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.LoginResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "nis",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nis": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handlers.RegisterResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_google_acct": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "nis": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}